#0: Cleanup falcon40b op tests
