+++ models/demos/t3000/falcon40b/tests/unit_tests/distributed_layernorm/distributed_layernorm_pytorch_references.py
