+++ models/demos/t3000/falcon40b/tests/unit_tests/test_falcon_create_qkv_heads.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+# SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
 
 # SPDX-License-Identifier: Apache-2.0
 
@@ -7,7 +7,6 @@ import pytest
 from loguru import logger
 
 import ttnn
-from models.demos.t3000.falcon40b.tt.ops.falcon_nlp_create_qkv_heads import TtFalconCreateQKVHeads
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import (
     comp_pcc,
 )
@@ -19,6 +18,8 @@ from models.demos.t3000.falcon40b.tt.model_config import (
 from models.demos.t3000.falcon40b.reference.hf_modeling_falcon import (
     FalconForCausalLM,
 )
+
+from typing import List
 
 
 class PytorchFalconCreateQKVHeads(torch.nn.Module):
@@ -48,24 +49,37 @@ class PytorchFalconCreateQKVHeads(torch.nn.Module):
         return query, key, value
 
 
+class TtFalconCreateQKVHeads:
+    def __init__(
+        self,
+        device,
+        num_heads: int = 32,
+        num_kv_heads: int = 2,
+    ):
+        super().__init__()
+
+        self.num_heads = num_heads
+        self.num_kv_heads = num_kv_heads
+
+    def __call__(self, x: ttnn.experimental.tensor.Tensor) -> ttnn.experimental.tensor.Tensor:
+        q_layer, k_layer, v_layer = ttnn.experimental.tensor.nlp_create_qkv_heads(
+            x,
+            num_heads=self.num_heads,
+            num_kv_heads=self.num_kv_heads,
+            transpose_k_heads=False,
+            output_mem_config=ttnn.MemoryConfig(ttnn.TensorMemoryLayout.INTERLEAVED, ttnn.BufferType.DRAM),
+        )
+
+        return q_layer, k_layer, v_layer
+
+
 def run_test_FalconMLP_inference(
     pcc,
     device,
-    model_location_generator,
 ):
-    # # mistral
-    # num_heads = 32
-    # num_kv_heads = 8
-    # head_dim = 128
-
     # falcon 40b per chip: 8 chip setup
     num_heads = 16
     num_kv_heads = 1
-    head_dim = 64
-
-    # falcon 40b per chip: 4 chip setup
-    num_heads = 32
-    num_kv_heads = 2
     head_dim = 64
 
     seqlen = 32
@@ -73,8 +87,6 @@ def run_test_FalconMLP_inference(
     # Prepare input
     torch.manual_seed(0)
     model_input_shape = [1, seqlen]
-
-    model_config = get_model_config("BFLOAT8_B-SHARDED", "prefill", model_input_shape, 4)
 
     input_shape = [
         1,
@@ -84,13 +96,6 @@ def run_test_FalconMLP_inference(
     ]  # 2304 = (16 + 1 + 1) * 2 groups * 64 head_dim
     input = (torch.rand(input_shape) * 2) - 1
 
-    model_version = "tiiuae/falcon-40b-instruct"
-    model_name = model_location_generator(model_version, model_subdir="Falcon")
-
-    hugging_face_reference_model = FalconForCausalLM.from_pretrained(
-        model_name, low_cpu_mem_usage=True, num_hidden_layers=1
-    )
-
     # PyTorch output --------------------------------------------------------------------
     pytorch_FalconCreateQKVHeads_model = PytorchFalconCreateQKVHeads(
         num_heads=num_heads, num_kv_heads=num_kv_heads, head_dim=head_dim
@@ -99,7 +104,9 @@ def run_test_FalconMLP_inference(
 
     # TT hardware execution -------------------------------------------------------------
     tt_Falconcreate_qkv_heads_model = TtFalconCreateQKVHeads(
-        device, model_config=model_config, num_heads=num_heads, num_kv_heads=num_kv_heads, head_dim=head_dim
+        device,
+        num_heads=num_heads,
+        num_kv_heads=num_kv_heads,
     )
 
     input_host = torch2tt_tensor(input, None, tt_dtype=ttnn.experimental.tensor.DataType.BFLOAT16)
@@ -142,12 +149,10 @@ def run_test_FalconMLP_inference(
 def test_FalconMatmul_inference(
     pcc,
     all_devices,
-    model_location_generator,
 ):
     device = all_devices[0]
 
     run_test_FalconMLP_inference(
         pcc,
         device,
-        model_location_generator,
     )
